test(front): source boundary editor state spec 분리
